requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
psycopg2-binary
pandas
numpy
python-dotenv

streamlit
plotly